+++ plante.h
@@ -1,0 +1,16 @@
+#ifndef Plante_h
+#define Plante_h
+
+typedef struct{
+    char id[25];
+    char ph[20];
+    char jord[20];
+    char fugt[10];
+    int hjort;
+    int hare;
+    int fugl;
+    int insekt;
+    int score;
+}Plante;
+
+#endif
