+++ util.h
@@ -1,0 +1,14 @@
+#ifndef Util_h
+#define Util_h
+
+#include "plante.h"
+
+char get_single_char();
+
+void clrscr();
+
+void insert_plant(Plante *p_arr, Plante p, int n);
+
+void print_array(Plante *p_arr, int n);
+
+#endif
